update grayscale basemap
